Remove unused code, move compiler to ViewCompiler namespace
