+++ src/Framework/Framework/Compilation/ControlTree/IAbstractTreeRoot.cs
@@ -1,5 +1,6 @@
 ﻿
 using System.Collections.Generic;
+using DotVVM.Framework.Compilation.ViewCompiler;
 
 namespace DotVVM.Framework.Compilation.ControlTree
 {
