+++ src/Framework/Framework/Compilation/IControlBuilderFactory.cs
@@ -1,4 +1,5 @@
 using System;
+using DotVVM.Framework.Compilation.ViewCompiler;
 
 namespace DotVVM.Framework.Compilation
 {
